API: Rename Chatserver.ts to ChatServer.ts
